+++ src/main/kotlin/lotto/domain/lotto/Lotto.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.lotto
 
 import lotto.exception.LottoException
