+++ src/main/kotlin/lotto/domain/lotto/LottoNumber.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.lotto
 
 import lotto.exception.LottoNumberException
