+++ src/test/kotlin/lotto/domain/lotto/LottoTest.kt
@@ -1,4 +1,4 @@
-package lotto.domain
+package lotto.domain.lotto
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
